Add PastEvents component and extend FutureEvents functionality
A new PastEvents component has been added to display past events in the system. Functionality has also been extended on the FutureEvents component to include the ability to add a new event. Additional modifications include styling updates on the HomeView, and the implementation of a new API service to fetch past events.