Add active link styles in navbar
The commit adds an active link style for the "Home" and "Add Event" links in the navigation bar of the application. This style change includes a bright background and white text color for the active link. Moreover, a smooth transition has been added to the color change of the navbar links.