Add NotFoundPage route and style adjustments
An additional NotFoundPage route has been added to the router file to handle any path not recognized by the app. The NotFoundPage displays an error message informing the user that the requested page does not exist and provides a link back to the homepage. Minimal padding-right has also been added to the App.vue styling for aesthetic adjustment.